+++ braindecode/datasets/tuh.py
@@ -45,6 +45,10 @@ class TUH(BaseConcatDataset):
     add_physician_reports: bool
         If True, the physician reports will be read from disk and added to the
         description.
+    rename_channels: bool
+        If True, rename the EEG channels to the standard 10-05 system.
+    set_montage: bool
+        If True, set the montage to the standard 10-05 system.
     n_jobs: int
         Number of jobs to be used to read files in parallel.
     """
@@ -56,8 +60,14 @@ class TUH(BaseConcatDataset):
         target_name: str | None = None,
         preload: bool = False,
         add_physician_reports: bool = False,
+        rename_channels: bool = False,
+        set_montage: bool = False,
         n_jobs: int = 1,
     ):
+        if set_montage:
+            assert (
+                rename_channels
+            ), "If set_montage is True, rename_channels must be True."
         # create an index of all files and gather easily accessible info
         # without actually touching the files
         file_paths = glob.glob(os.path.join(path, "**/*.edf"), recursive=True)
@@ -78,26 +88,92 @@ class TUH(BaseConcatDataset):
         if n_jobs == 1:
             base_datasets = [
                 self._create_dataset(
-                    descriptions[i], target_name, preload, add_physician_reports
+                    descriptions[i],
+                    target_name,
+                    preload,
+                    add_physician_reports,
+                    rename_channels,
+                    set_montage,
                 )
                 for i in descriptions.columns
             ]
         else:
             base_datasets = Parallel(n_jobs)(
                 delayed(self._create_dataset)(
-                    descriptions[i], target_name, preload, add_physician_reports
+                    descriptions[i],
+                    target_name,
+                    preload,
+                    add_physician_reports,
+                    rename_channels,
+                    set_montage,
                 )
                 for i in descriptions.columns
             )
         super().__init__(base_datasets)
 
     @staticmethod
-    def _create_dataset(description, target_name, preload, add_physician_reports):
+    def _rename_channels(raw):
+        """
+        Renames the EEG channels using mne conventions and sets their type to 'eeg'.
+
+        See https://isip.piconepress.com/publications/reports/2020/tuh_eeg/electrodes/
+        """
+        # remove ref suffix and prefix:
+        # TODO: replace with removesuffix and removeprefix when 3.8 is dropped
+        mapping_strip = {
+            c: c.replace("-REF", "").replace("-LE", "").replace("EEG ", "")
+            for c in raw.ch_names
+        }
+        raw.rename_channels(mapping_strip)
+
+        montage1005 = mne.channels.make_standard_montage("standard_1005")
+        mapping_eeg_names = {
+            c.upper(): c for c in montage1005.ch_names if c.upper() in raw.ch_names
+        }
+
+        # Set channels whose type could not be inferred (defaulted to "eeg") to "misc":
+        non_eeg_names = [c for c in raw.ch_names if c not in mapping_eeg_names]
+        if non_eeg_names:
+            non_eeg_types = raw.get_channel_types(picks=non_eeg_names)
+            mapping_non_eeg_types = {
+                c: "misc" for c, t in zip(non_eeg_names, non_eeg_types) if t == "eeg"
+            }
+            if mapping_non_eeg_types:
+                raw.set_channel_types(mapping_non_eeg_types)
+
+        if mapping_eeg_names:
+            # Set 1005 channels type to "eeg":
+            raw.set_channel_types(
+                {c: "eeg" for c in mapping_eeg_names}, on_unit_change="ignore"
+            )
+            # Fix capitalized EEG channel names:
+            raw.rename_channels(mapping_eeg_names)
+
+    @staticmethod
+    def _set_montage(raw):
+        montage = mne.channels.make_standard_montage("standard_1005")
+        raw.set_montage(montage, on_missing="ignore")
+
+    @staticmethod
+    def _create_dataset(
+        description,
+        target_name,
+        preload,
+        add_physician_reports,
+        rename_channels,
+        set_montage,
+    ):
         file_path = description.loc["path"]
 
         # parse age and gender information from EDF header
         age, gender = _parse_age_and_gender_from_edf_header(file_path)
-        raw = mne.io.read_raw_edf(file_path, preload=preload)
+        raw = mne.io.read_raw_edf(
+            file_path, preload=preload, infer_types=True, verbose="error"
+        )
+        if rename_channels:
+            TUH._rename_channels(raw)
+        if set_montage:
+            TUH._set_montage(raw)
 
         meas_date = (
             datetime(1, 1, 1, tzinfo=timezone.utc)
@@ -312,6 +388,12 @@ class TUHAbnormal(TUH):
     add_physician_reports: bool
         If True, the physician reports will be read from disk and added to the
         description.
+    rename_channels: bool
+        If True, rename the EEG channels to the standard 10-05 system.
+    set_montage: bool
+        If True, set the montage to the standard 10-05 system.
+    n_jobs: int
+        Number of jobs to be used to read files in parallel.
     """
 
     def __init__(
@@ -321,6 +403,8 @@ class TUHAbnormal(TUH):
         target_name: str | None = "pathological",
         preload: bool = False,
         add_physician_reports: bool = False,
+        rename_channels: bool = False,
+        set_montage: bool = False,
         n_jobs: int = 1,
     ):
         with warnings.catch_warnings():
@@ -333,6 +417,8 @@ class TUHAbnormal(TUH):
                 preload=preload,
                 target_name=target_name,
                 add_physician_reports=add_physician_reports,
+                rename_channels=rename_channels,
+                set_montage=set_montage,
                 n_jobs=n_jobs,
             )
         additional_descriptions = []
@@ -442,6 +528,8 @@ class _TUHMock(TUH):
         target_name: str | None = None,
         preload: bool = False,
         add_physician_reports: bool = False,
+        rename_channels: bool = False,
+        set_montage: bool = False,
         n_jobs: int = 1,
     ):
         with warnings.catch_warnings():
@@ -452,6 +540,8 @@ class _TUHMock(TUH):
                 target_name=target_name,
                 preload=preload,
                 add_physician_reports=add_physician_reports,
+                rename_channels=rename_channels,
+                set_montage=set_montage,
                 n_jobs=n_jobs,
             )
 
@@ -474,6 +564,8 @@ class _TUHAbnormalMock(TUHAbnormal):
         target_name: str | None = "pathological",
         preload: bool = False,
         add_physician_reports: bool = False,
+        rename_channels: bool = False,
+        set_montage: bool = False,
         n_jobs: int = 1,
     ):
         with warnings.catch_warnings():
@@ -484,5 +576,7 @@ class _TUHAbnormalMock(TUHAbnormal):
                 target_name=target_name,
                 preload=preload,
                 add_physician_reports=add_physician_reports,
+                rename_channels=rename_channels,
+                set_montage=set_montage,
                 n_jobs=n_jobs,
             )