+++ docs/whats_new.rst
@@ -34,6 +34,7 @@ Enhancements
 - Increasing the coverage of the tests (:gh:`592` by `Bruno Aristimunha`_)
 - Adding cache and pre-processing option to :class:`braindecode.datasets.MOABBDataset` (:gh:`582` by `Bruno Aristimunha`_)
 - Add type hints to datasets (:gh:`590` by `Pierre Guetschel`_)
+- Add channel names and montage to  :class:`braindecode.datasets.TUH` (:gh:`593` by `Pierre Guetschel`_)
 
 
 Bugs